some fixes in SingleQueryService

+++ src/main/java/ru/dawgg/reportbuilderapi/service/impl/SingleQueryServiceImpl.java
@@ -3,11 +3,9 @@ package ru.dawgg.reportbuilderapi.service.impl;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import org.springframework.data.util.Streamable;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
-import ru.dawgg.reportbuilderapi.exception.InvalidSingleQueryIdException;
+import ru.dawgg.reportbuilderapi.exception.SingleQueryNotFoundException;
 import ru.dawgg.reportbuilderapi.model.query.SingleQuery;
 import ru.dawgg.reportbuilderapi.repository.SingleQueryRepository;
 import ru.dawgg.reportbuilderapi.service.SingleQueryService;
@@ -22,43 +20,57 @@ public class SingleQueryServiceImpl implements SingleQueryService {
     private final JdbcTemplate template;
 
     @Override
-    public ResponseEntity<Object> save(SingleQuery singleQuery) {
-        repository.save(singleQuery);
-
-        return new ResponseEntity<>(HttpStatus.CREATED);
+    public void save(SingleQuery singleQuery) {
+        if (isExistOrThrow(singleQuery)) {
+            repository.save(singleQuery);
+        }
     }
 
     @Override
-    public ResponseEntity<Object> update(SingleQuery singleQuery) {
-        save(singleQuery);
-
-        return ResponseEntity.ok(singleQuery);
+    public void update(SingleQuery singleQuery) {
+        if (isExistOrThrow(singleQuery)) {
+            save(singleQuery);
+        }
     }
 
     @Override
-    public ResponseEntity<Object> deleteById(Integer id) {
-        repository.deleteById(id);
-
-        return new ResponseEntity<>(HttpStatus.ACCEPTED);
+    public void deleteById(Integer id) {
+        if (isExistByIdOrThrow(id)) {
+            repository.deleteById(id);
+        }
     }
 
     @Override
-    public ResponseEntity<Object> executeById(Integer id) {
-        var singleQuery = findById(id);
-        template.execute(singleQuery.getQuery());
-
-        return new ResponseEntity<>(HttpStatus.CREATED);
+    public void executeById(Integer id) {
+        if (isExistByIdOrThrow(id)) {
+            var singleQuery = findById(id);
+            template.execute(singleQuery.getQuery());
+        }
     }
 
     @Override
     @SneakyThrows
     public SingleQuery findById(Integer id) {
         return repository.findById(id)
-                .orElseThrow(() -> new InvalidSingleQueryIdException("Query with the id has not been found"));
+                .orElseThrow(() -> new SingleQueryNotFoundException("Query with the id has not been found"));
     }
 
     @Override
     public List<SingleQuery> findAll() {
         return Streamable.of(repository.findAll()).toList();
+    }
+
+    @SneakyThrows
+    private boolean isExistOrThrow(SingleQuery singleQuery) {
+        if (repository.exists(singleQuery)) {
+            return true;
+        } else throw new SingleQueryNotFoundException("The query has not been found");
+    }
+
+    @SneakyThrows
+    private boolean isExistByIdOrThrow(Integer id) {
+        if (repository.existsById(id)) {
+            return true;
+        } else throw new SingleQueryNotFoundException("Query with id - " + id + " has not been found");
     }
 }
